fdo#88475 RowSetBase: reposition cache before interrogating it
This partially reverts:

  commit 48d98be318430f5fdd30c6aa2b5210d791ee629d
  Author: Lionel Elie Mamane <lionel@mamane.lu>
  Date:   Thu Dec 6 13:22:06 2012 +0100

      Don't force refresh when higher up code did not request it

      Change-Id: I0f415c96fc05c1d776d14885751aef020c42f4ae

which did not take into account that the cache is shared with the
clones.

Finish the above commit more deeply, that is
more aggressively avoid moving the cache if we don't need to:
when we merely reposition the cache in case another RowSet(Clone)
has moved it, no need to force a move and by that to force a refresh.

Change-Id: I28d62673fdf10ee6507d38bb7c79c08e4b40902f