On init of DbCellControl, load value from Model
This fixes a regression introduced by "fdo#88551 no need to update content from field when activating grid cell"
that when switching from design mode to normal mode
before one moved row at least once, the current cell (inevitably of the first row)
was always displayed empty.
Before, the update was done by the activation event (which was too often), so now we need to do it once at initialisation.
Not sure if using updateFromModel instead of UpdateFromField makes a difference. Hope not. Else need to change that.

Change-Id: I878dea0f91b370a4f83c3c1a3ed185a51ac9f0b7