Revert "fdo#88475 BrowseBox/grid: reposition data cursor to current row after paint"
(and its later fixup)

This reverts commits
 b70d6df94078ee7969cd44c2841bb797497d8d93
 9723460aae4a3ad69e00e1f22c19461f583c9913

It was a work-around for the real underlying issue, which was that the
result of dbaccess::ResultSet::isFirst() et al were clobbered by moves
made by its clones.

The BrowseBox has two different cursors:
1) One for data to edit (which is kept on the current/active row)
2) One for data to *paint*
The second is a clone of the first.

The real underlying issue is fixed by:

  commit a046063d129e52ec5915f6a96647112c0d7c978e
  Author: Lionel Elie Mamane <lionel@mamane.lu>
  Date:   Thu Jan 22 10:49:42 2015 +0100

      fdo#88475 RowSetBase: reposition cache before interrogating it

      Change-Id: I28d62673fdf10ee6507d38bb7c79c08e4b40902f