Resolves: tdf#87083 images in help html stuck at default placeholder size
regression from the SwapIn work.

The graphic size during SetGrfFlySize appears to assume its the pre-swapped in
size, i.e. the old 0 by 0 "size is not set" for an html image used to be the
value used here. So get that value (and update it if necessary) before updating
the image data and pass that to SetGrfFlySize

Without the swap-in in the html filter the data gets swapped in during the
initial draw of the image which is too late to update the image size, so
restore the equivalent of that swap-in via GetTwipSize

commit 7ee209935d0326c102135639f5af2491078d42dc
Author: Zolnai Tamás <tamas.zolnai@collabora.com>
Date:   Fri Nov 7 10:43:34 2014 +0100

    Make SwGrfNode swapping methods private

    Change the filter test accordingly.

    Change-Id: Ide3043f2f245c097a7b4c07ba2e0713510296b3e

Change-Id: I0afe244f79d8628236b1e552036587a2b4540a76