SwSelPaintRects: add FillStartEnd() interface
Previously we always assumed non-table selections in SwSelPaintRects,
but that is not always true, resulting in incorrect handles for table
selections. Add a FillStartEnd() interface to fill the passed SwRects
with the rectangle of the start and end handle, and implement it in both
SwShellCrsr and SwShellTableCrsr.

This makes adjusting the end of table selections using the end selection
handle work (be it shrinking or extending).

Change-Id: Iba8657466b102ec60c41829f00660dfe295643ab