gbuild: CppunitTest: always use unittest configuration
The sc_subsequent_filters_test was failing because of a lock file
because it did not use the unittest configuration.

Refactor gb_CppunitTest_use_configuration so it uses both the instdir
and unittest configuration to prevent such errors.

In case there ever is a test that does not work with the unittest
configuration it should call gb_CppunitTest_use_instdir_configuration.

Change-Id: Ibc00d42f8b6102d50d922f51173120798fa45c6e