SwCursor: when tiled rendering also select word before the point
On the desktop, the logic is that if we double-click before a word, then we
select the word, but if we double-click after it, then we don't.

When tiled rendering, it's more important to always select something, as
that's how the selection can be adjusted further. So if we long push
after a non-empty text node, then select its last word. This way
something is always selected if e.g. the user long pushes inside a table
cell. (If the table cell is empty, then that's handled in
SwCrsrShell::SelectCell().)

Change-Id: I6e84874b8cf252fc6b72a75ffe9db67b03636862